fix circular dependency
